services/consumer/requirements.txt
kafka-python==2.0.2
fastavro==1.8.0
pydantic==1.10.12
psycopg2-binary==2.9.10
python-dotenv==1.0.0
